added Permutations Solution

+++ IntroductoryProblems/IncreasingArray.cpp
@@ -9,7 +9,7 @@ void solve() {
    for(int i = 0; i < n; i++){
         cin >> arr[i];
    }
-    long long cnt = 0;
+    int cnt = 0;
     for(int i = 1; i < n; i++){
         if(arr[i] < arr[i-1]){
             cnt += arr[i-1] - arr[i];
